Patch #5234: Fix incorrect type in GaudiAlg tuples
https://savannah.cern.ch/patch/?5234

From: Christopher R Jones <christopher.rob.jones@cern.ch>


git-svn-id: svn+ssh://svn.cern.ch/reps/gaudi/Gaudi/trunk@6823 53ecefc3-a64d-0410-9bdf-b8581fa3287e

+++ GaudiAlg/GaudiAlg/TupleObj.h
@@ -708,7 +708,7 @@ namespace Tuples
      *  @return status code
      */
     StatusCode column ( const std::string&       name  ,
-                        const int                value ,
+                        const unsigned long long value ,
                         const unsigned long long minv  ,
                         const unsigned long long maxv  );
     // ========================================================================
